[2단계 - 장바구니 미션] 렛서(김다은) 미션 제출합니다. (#314)
* chore: css LayoutWrapper에서 height 설정 제거

* chore: Layout의 Content 컨테이너 스타일 변경

* refactor: Header의 버튼을 prop으로 전달하도록 변경

* refactor: 주문 확인 페이지를 결제 확인 페이지로 변경 및 그에 따른 세부 변경사항

- 이름 변경(ConfirmOrderPage -> CheckoutPage)
- 라우팅 변경('/confirm' -> '/checkout')
- BottomBotton의 prop 이름을 isDisable -> active으로 변경하고, 관련 로직 수정
- 기타 스타일 수정

* refactor: CartItemQuantity 리팩토링

- summary 옵션 추가
- 컴포넌트 분리
- 수량 관련 로직을 useQuantityControls로 추상화

* chore: CartList 컴포넌트 스타일 변경

* chore: Layout 스타일 수정

* refactor: recoil 파일 분리

* refactor: recoil import 경로 수정

* rerfactor: 도서 산간 지역 배송비 선택 상태 추가, 배송비 관련 상태를 shipping.ts로 분리

* refactor: Page 하위의 commonStyles를 components 하위의 commonStyles로 이동

* feat: 아무것도 선택하지 않으면 주문 확인 페이지로 넘어갈 수 없도록 변경

* feat: 배송 옵션 선택 컴포넌트 추가

* refactor: api, type 정의 파일을 관심사에 맞게 분리

* feat: 쿠폰 선택 모달 UI

* refactor: 반복되는 toLacaleString 사용을 유틸 함수로 분리

* fix: 삭제 api url 링크 수정

* feat: 삭제하기 전 사용자 의사를 한 번 더 묻도록 수정

* refactor: 쿠폰 적용 버튼의 클릭 핸들러 분리

* chore: CartList 및 CartItem 너비, 간격 관련 스타일 수정

* feat: 주문 확인 페이지 UI 추가

* feat: 결제 확인 페이지(CheckoutPage)에서 Header의 버튼 제거

* refactor: 주문 정보를 selector에서 받아와 객체로 반환하는 useOrderInformation 훅 생성

* feaat: 쿠폰 선택 기능 추가

* feat: 할인 금액 계산 기능 추가

* feat: 주문 post 요청 기능 추가

* fix: 장바구니로 돌아가기 버튼을 누르면 홈으로 라우팅 하는 게 아닌 새로고침을 하도록 수정하여 장바구니 리스트를 refetch

* fix: BOGO 쿠폰 적용 가능 여부 판정 버그

* chore: 불필요한 주석 제거

* refactor: 테스트 코드의 selector, atom 파일 import 수정

* chore: 사용하지 않는 import 제거

* chore: storybook 컴포넌트 import 경로 수정

* refactor: window.location.href 설정 대신 navigate 사용 후 reload하도록 변경

* feat: 서버에 데이터 추가 요청 후 페이지를 새로고침 하도록 변경

* chore: jest 설정 관련 파일 제거

* fix: BOGO 쿠폰에서 참조하는 selectedItems의 quantity가 업데이트 되지 않는 버그 수정

* chore: 구조 분해 할당 시 이름을 부여하는 것으로 함수 제거

* refactor: timeFormatter의 매개변수로 전달되는 문자열의 타입 지정

* chore: 불필요한 주석 제거

* chore: 잘못 작성된 async, await 수정

* refactor: commonStyles 위치 변경

+++ src/hooks/useDeleteCartItem.ts
@@ -1,21 +1,25 @@
 import { startTransition } from 'react';
 import { useRecoilRefresher_UNSTABLE } from 'recoil';
 import { deleteCartItem } from '../api';
-import { cartListState } from '../recoil/selectors';
+import { cartListSelector } from '../recoil';
 import CartItemLocalStorage, {
   CART_ITEM_SELECTED_KEY,
 } from '../services/CartItemLocalStorage';
 
 export default function useDeleteCartItem(id: number) {
-  const refreshCartList = useRecoilRefresher_UNSTABLE(cartListState);
+  const refreshCartList = useRecoilRefresher_UNSTABLE(cartListSelector);
 
   const handleDelete = async () => {
-    await deleteCartItem(id);
-    CartItemLocalStorage.delete(CART_ITEM_SELECTED_KEY, id);
+    const confirm = window.confirm('선택한 제품을 삭제하겠습니까?');
 
-    startTransition(() => {
-      refreshCartList();
-    });
+    if (confirm) {
+      await deleteCartItem(id);
+      CartItemLocalStorage.delete(CART_ITEM_SELECTED_KEY, id);
+
+      startTransition(() => {
+        refreshCartList();
+      });
+    }
   };
 
   return { handleDelete };
